Add files via upload
This is an assembler for HACK computer written in node for the Nand2Tetris course week 6. I didn't add the symbol and variables translation, but it can translate all the basic assembly language code into HACK machine language. To test it, just type "node main.js" in the console and it will output the binary file in out.txt.